GH-78: start implementing lobby based deployment
Initial implementation uses both network and shared memory channels for message processing. Agents network channel forwards messages to shared memory which are processed by shared memory channel.
The following improvements were implemented in the core dds-protocol:
dds-protocol: Added: new API for pushing and processing of raw messages. Implemented for network and shared memory channels.
dds-protocol: Added: special command ECmdType::cmdRAW_MSG  for raw message event subscription.
dds-protocol: Modified: create new message in the channel instead of clearing the current message. This allows to forward the message without additional copying.
