+++ dds-protocol-lib/src/ChannelMessageHandlersImpl.h
@@ -20,6 +20,7 @@ namespace dds
         class CChannelMessageHandlersImpl : private CBaseEventHandlersImpl<ECmdType>
         {
             DDS_BEGIN_EVENT_HANDLERS(ECmdType)
+            DDS_REGISTER_EVENT_HANDLER(ECmdType, cmdRAW_MSG, void(protocol_api::CProtocolMessage::protocolMessagePtr_t))
             DDS_REGISTER_MESSAGE_HANDLER(cmdREPLY_HANDSHAKE_OK)
             DDS_REGISTER_MESSAGE_HANDLER(cmdREPLY_HANDSHAKE_ERR)
             DDS_REGISTER_MESSAGE_HANDLER(cmdSHUTDOWN)
